Finally works

+++ include/integral.h
@@ -1,7 +1,7 @@
 #ifndef __INTEGRAL_H__
 #define __INTEGRAL_H__
 
-void freqGpuSimulate(double *, double *, double *, double *, int, double *, int, double, double, int); 
+void freqGpuSimulate(float *, float *, float *, float *, int, float *, int, float, float, int); 
 
 #endif /* __INTEGRAL_H__ */
